+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/importers/SwfXmlImporter.java
@@ -37,6 +37,7 @@ import com.jpexs.decompiler.flash.abc.types.traits.TraitMethodGetterSetter;
 import com.jpexs.decompiler.flash.abc.types.traits.TraitSlotConst;
 import com.jpexs.decompiler.flash.abc.types.traits.Traits;
 import com.jpexs.decompiler.flash.amf.amf3.Amf3Value;
+import com.jpexs.decompiler.flash.tags.CSMSettingsTag;
 import com.jpexs.decompiler.flash.tags.DefineSpriteTag;
 import com.jpexs.decompiler.flash.tags.Tag;
 import com.jpexs.decompiler.flash.tags.TagTypeInfo;
@@ -485,6 +486,10 @@ public class SwfXmlImporter {
     private Object createObject(String type, int tagTypeId, SWF swf, Tag tag) throws NoSuchMethodException, InstantiationException, IllegalAccessException, IllegalArgumentException, InvocationTargetException {
         if ("UnknownTag".equals(type)) {
             return new UnknownTag(swf, tagTypeId);
+        }
+        
+        if ("CSMTextSettings".equals(type)) {
+            type = CSMSettingsTag.NAME;
         }
 
         Class cls = swfTags.get(type);